tools/requirements.txt
# Core dependencies (required)
numpy>=1.21.0
psutil>=5.8.0

# GPU dependencies (installed only when available)
# CUDA 12.x+ cards (Linux/Windows):
cupy-cuda12x>=12.0.0; platform_system == "Linux"

# Apple Silicon (Metal Performance Shaders):
# torch>=2.0.0; platform_system == "Darwin"

# AMD ROCm GPUs:
# cupy-rocm-5-0>=12.0.0; platform_system == "Linux"

